Extract submitter supplied protocol into it's own file
